+++ ros_ethercat_eml/include/ros_ethercat_eml/ethercat_slave_handler.h
@@ -40,6 +40,7 @@
 class EtherCAT_AL;
 
 /// EtherCAT Slave Handler
+
 class EtherCAT_SlaveHandler : public EC_ESM, public EtherCAT_SlaveConfig
 {
   // Being friendly
@@ -68,11 +69,17 @@ public:
                         EC_FixedStationAddress a_station_address,
                         EtherCAT_FMMU_Config * a_fmmu_config,
                         EtherCAT_PD_Config * a_pd_config,
-                        EtherCAT_MbxConfig * a_mbx_config = NULL);
+                        EtherCAT_MbxConfig * a_mbx_config,
+                        EtherCAT_DataLinkLayer *_m_dll_instance,
+                        EC_Logic *_m_logic_instance,
+                        EtherCAT_PD_Buffer *_m_pdbuf_instance);
   /// Constructor using Slave Configuration
   EtherCAT_SlaveHandler(uint16_t a_ring_position,
-                        const EtherCAT_SlaveConfig * a_sconf,
-                        uint32_t a_serial);
+                        EtherCAT_SlaveConfig * a_sconf,
+                        uint32_t a_serial,
+                        EtherCAT_DataLinkLayer *_m_dll_instance,
+                        EC_Logic *_m_logic_instance,
+                        EtherCAT_PD_Buffer *_m_pdbuf_instance);
 
   /// Get position in the EtherCAT logical ring
   uint16_t get_ring_position() const
@@ -87,7 +94,7 @@ public:
 
   /// Returns and increments sequence number used for duplication mailbox write dectition
   uint8_t get_mbx_counter();
-  protected:
+protected:
 
   /// Position in the EtherCAT logical Ring
   uint16_t m_ring_position;