+++ ros_ethercat_model/include/ros_ethercat_model/ros_ethercat.hpp
@@ -40,6 +40,8 @@
 #ifndef ROS_ETHERCAT_MODEL_ROS_ETHERCAT_HPP_
 #define ROS_ETHERCAT_MODEL_ROS_ETHERCAT_HPP_
 
+#include <boost/algorithm/string/split.hpp>
+#include <boost/algorithm/string.hpp>
 #include <ros/ros.h>
 #include <ros/console.h>
 #include <hardware_interface/robot_hw.h>
@@ -70,41 +72,55 @@
  */
 
 using std::string;
-using boost::unordered_map;
+using std::vector;
 using boost::ptr_unordered_map;
-using boost::scoped_ptr;
-using ros_ethercat_model::JointState;
-using hardware_interface::JointStateHandle;
-using hardware_interface::JointHandle;
+using boost::ptr_vector;
 using ros_ethercat_model::JointState;
 using ros_ethercat_model::Actuator;
-using ros_ethercat_model::RobotState;
-using ros::NodeHandle;
+using ros_ethercat_model::Transmission;
+using ros_ethercat_model::CustomHW;
 
 static const string name = "ros_ethercat";
 
 class RosEthercat : public hardware_interface::RobotHW
 {
 public:
-  RosEthercat(NodeHandle &nh, const string &eth, bool allow, TiXmlElement* config) :
+  RosEthercat(ros::NodeHandle &nh, const string &eth, bool allow, TiXmlElement* config) :
     cm_node_(nh, "ethercat_controller_manager"),
-    model_(config),
-    ec_(name, static_cast<hardware_interface::HardwareInterface*>(&model_), eth, allow)
+    model_(config)
   {
-    unordered_map<string, JointState>::iterator it = model_.joint_states_.begin();
-    while (it != model_.joint_states_.end())
+    vector<string> port_names;
+    boost::split(port_names, eth, boost::is_any_of("_, "));
+    for (vector<string>::const_iterator port_name = port_names.begin();
+         port_name != port_names.end();
+         ++port_name)
     {
-      JointStateHandle jsh(it->first,
-                           &it->second.position_,
-                           &it->second.velocity_,
-                           &it->second.measured_effort_);
+      if (!port_name->empty())
+      {
+        ethercat_hardware_.push_back(new EthercatHardware(name,
+                                                          static_cast<hardware_interface::HardwareInterface*> (&model_),
+                                                          *port_name,
+                                                          allow));
+        ROS_INFO_STREAM("Added Ethernet port " << *port_name);
+      }
+    }
+
+    for (ptr_unordered_map<string, JointState>::iterator it = model_.joint_states_.begin();
+         it != model_.joint_states_.end();
+         ++it)
+    {
+      hardware_interface::JointStateHandle jsh(it->first,
+                                               &it->second->position_,
+                                               &it->second->velocity_,
+                                               &it->second->effort_);
       joint_state_interface_.registerHandle(jsh);
 
-      JointHandle jh(joint_state_interface_.getHandle(it->first), &it->second.commanded_effort_);
-      joint_command_interface_.registerHandle(jh);
-      effort_joint_interface_.registerHandle(jh);
-      position_joint_interface_.registerHandle(jh);
-      ++it;
+      joint_position_command_interface_.registerHandle(hardware_interface::JointHandle(jsh,
+                                                                                       & it->second->commanded_position_));
+      joint_velocity_command_interface_.registerHandle(hardware_interface::JointHandle(jsh,
+                                                                                       & it->second->commanded_velocity_));
+      joint_effort_command_interface_.registerHandle(hardware_interface::JointHandle(jsh,
+                                                                                     & it->second->commanded_effort_));
     }
 
     if (!model_.joint_states_.empty())
@@ -112,64 +128,99 @@ public:
 
     registerInterface(&model_);
     registerInterface(&joint_state_interface_);
-    registerInterface(&joint_command_interface_);
-    registerInterface(&effort_joint_interface_);
-    registerInterface(&position_joint_interface_);
+    registerInterface(&joint_position_command_interface_);
+    registerInterface(&joint_velocity_command_interface_);
+    registerInterface(&joint_effort_command_interface_);
   }
 
-  virtual ~RosEthercat() {}
+  virtual ~RosEthercat()
+  {
+  }
 
   /// propagate position actuator -> joint and set commands to zero
-  void read()
+  void read(const ros::Time &time)
   {
-    ec_.update(false, false);
-    model_.current_time_ = ros::Time::now();
+    for (ptr_vector<EthercatHardware>::iterator eh = ethercat_hardware_.begin();
+         eh != ethercat_hardware_.end();
+         ++eh)
+    {
+      eh->update(false, false);
+    }
+
+    model_.current_time_ = time;
     model_.propagateActuatorPositionToJointPosition();
 
-    unordered_map<string, JointState>::iterator it = model_.joint_states_.begin();
-    while (it != model_.joint_states_.end())
+    for (ptr_unordered_map<std::string, CustomHW>::iterator it = model_.custom_hws_.begin();
+         it != model_.custom_hws_.end();
+         ++it)
     {
-      it->second.joint_statistics_.update(&it->second);
-      it->second.commanded_effort_ = 0;
-      ++it;
+      it->second->read(time);
+    }
+
+    for (ptr_unordered_map<string, JointState>::iterator it = model_.joint_states_.begin();
+         it != model_.joint_states_.end();
+         ++it)
+    {
+      it->second->joint_statistics_.update(it->second);
+      it->second->commanded_effort_ = 0;
     }
   }
 
   /// propagate effort joint -> actuator and enforce safety limits
-  void write()
+  void write(const ros::Time &time)
   {
     /// Modify the commanded_effort_ of each joint state so that the joint limits are satisfied
-    unordered_map<string, JointState>::iterator it = model_.joint_states_.begin();
-    while (it != model_.joint_states_.end())
+    for (ptr_unordered_map<string, JointState>::iterator it = model_.joint_states_.begin();
+         it != model_.joint_states_.end();
+         ++it)
     {
-      it->second.enforceLimits();
-      ++it;
+      it->second->enforceLimits();
     }
+
     model_.propagateJointEffortToActuatorEffort();
+
+    for (ptr_unordered_map<std::string, CustomHW>::iterator it = model_.custom_hws_.begin();
+         it != model_.custom_hws_.end();
+         ++it)
+    {
+      it->second->write(time);
+    }
+
     if (!model_.joint_states_.empty())
-      mech_stats_publisher_->publish();
+      mech_stats_publisher_->publish(time);
   }
 
+  /// stop all actuators
   void shutdown()
   {
-    ptr_unordered_map<string, Actuator>::iterator it = model_.actuators_.begin();
-    while (it != model_.actuators_.end())
+    for (ptr_vector<Transmission>::iterator it = model_.transmissions_.begin();
+         it != model_.transmissions_.end();
+         ++it)
     {
-      it->second->command_.enable_ = false;
-      it->second->command_.effort_ = 0;
-      ++it;
+      it->actuator_->command_.enable_ = false;
+      it->actuator_->command_.effort_ = 0;
+    }
+
+    for (ptr_vector<EthercatHardware>::iterator eh = ethercat_hardware_.begin();
+         eh != ethercat_hardware_.end();
+         ++eh)
+    {
+      eh->update(false, true);
     }
   }
 
-  NodeHandle cm_node_;
-  RobotState model_;
-  EthercatHardware ec_;
-  scoped_ptr<MechStatsPublisher> mech_stats_publisher_;
+  ros::NodeHandle cm_node_;
+  ros_ethercat_model::RobotState model_;
+  ptr_vector<EthercatHardware> ethercat_hardware_;
+  boost::scoped_ptr<MechStatsPublisher> mech_stats_publisher_;
 
+  // state interface
   hardware_interface::JointStateInterface joint_state_interface_;
-  hardware_interface::JointCommandInterface joint_command_interface_;
-  hardware_interface::EffortJointInterface effort_joint_interface_;
-  hardware_interface::PositionJointInterface position_joint_interface_;
+
+  // command interface
+  hardware_interface::PositionJointInterface joint_position_command_interface_;
+  hardware_interface::VelocityJointInterface joint_velocity_command_interface_;
+  hardware_interface::EffortJointInterface joint_effort_command_interface_;
 };
 
 #endif